Improve regexp match error messages
Previously, if a regexp didn’t match at all the fault returned by
`expect-regexp-match` would simply say that it expected not false in
the results of `regexp-match` and instead found true. With this change,
first the expectation checks the regexp matches at all and uses an
attribute in the fault if it doesn’t, and only if it does match is the
result expectation used with a regexp match context added to its faults.